Start typescript migration for section js files
