+++ csm_web/frontend/src/components/Course.js
@@ -10,8 +10,8 @@ import ClockIcon from "../../static/frontend/img/clock.svg";
 import CheckCircle from "../../static/frontend/img/check_circle.svg";
 import XCircle from "../../static/frontend/img/x_circle.svg";
 import Modal, { ModalCloser } from "./Modal";
-import { SPACETIME_SHAPE } from "../utils/types";
-import { DAYS_OF_WEEK } from "./MentorSection";
+import { SPACETIME_SHAPE } from "../utils/js_types";
+import { DAYS_OF_WEEK } from "./section/utils";
 import TimeInput from "./TimeInput";
 
 const DAY_OF_WEEK_ABREVIATIONS = Object.freeze({
