+++ csm_web/frontend/src/utils/js_types.js
